requirements.txt
opencv-python>=4.5.4
mediapipe-silicon
numpy>=1.21.0 